requirements.txt
apex==0.9.10dev
diffusers==0.35.2
gradio==6.2.0
huggingface_hub==0.36.0
imageio==2.37.2
lora==0.3.0
matplotlib==3.10.8
model==0.6.0
numpy==2.4.0
pandas==2.3.3
peft==0.18.0
Pillow==12.1.0
scikit_learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
tensorboardx==2.6.4
thop==0.1.1.post2209072238
timm==1.0.24
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
transformers==4.57.3
wandb==0.23.1
